act2 & settings.json

+++ pertemuan 4/act2.cpp
@@ -2,17 +2,17 @@
 
 using namespace std;
 
-float luasSegitiga(int alas, int tinggi){
+float luasSegitiga(float alas, float tinggi){
     return tinggi*alas/2;
 }
 
-float kelilingSegitiga(int sisi1, int sisi2, int sisi3){
+float kelilingSegitiga(float sisi1, float sisi2, float sisi3){
     return sisi1+sisi2+sisi3;
 }
 
 int main()
 {
     cout << "Luas segitiga      : " << luasSegitiga(3, 5) << endl;
-    cout << "Keliling segitiga  : " <<kelilingSegitiga(3, 3, 4) << endl;
+    cout << "Keliling segitiga  : " << kelilingSegitiga(3, 3, 4) << endl;
     return 0;
 }
